msm: kgsl: Look for freed entries with the pagetable name
The list of freed entries stores the pagetable name which is usually
the process id. Search through this list using the process id and not
the thread id so that we see the entry we're interested in.

Change-Id: I4b19d8f47be0973b4f1ecf7dc2641b1824c6ca56
Signed-off-by: Lynus Vaz <lvaz@codeaurora.org>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>
(cherry picked from commit 3c180efb9e706ab29f30ebc2cd21464f12011ffc)