ANDROID: cpufreq: change uid_cpupower to use u32
The uid_cpupower files use u64 to report how long uid's have spent
on the cpu. Using u64 instead of u32 has a noticable impact on
the time taken to read a file. These file reads are on a critical
path to turning the screen on.

Using u64:
/proc/uid_cpupower/time_in_state
0.00584s real
0.0000134s user
0.00579s system

/proc/uid_cpupower/concurrent_active_time
0.00243s real
0.00000335s user
0.00241s system

/proc/uid_cpupower/concurrent_policy_time
0.00244s real
0.00000320s user
0.00243s system

Using u32:
/proc/uid_cpupower/time_in_state
0.00358s real
0.00000488s user
0.00356s system

/proc/uid_cpupower/concurrent_active_time
0.000851s real
0.00000274s user
0.000843s system

/proc/uid_cpupower/concurrent_policy_time
0.000824s real
0.00000244s user
0.000817s system

When using u32 the max time that can be reported is 497 days.
Since it is unlikely that a uid has run for that long without
the phone rebooting or the uid dying, the performance gains make
switching to u32 worth it. In the worst case, the non-essential
battery blaming will be inaccurate for that uid.

This patch does not change the internal storage size or how other
related stats are reported. It only changes how uid_cpupower
reports the times.

b/72763654
Test: measure read times and verify file contents is correct

Change-Id: I461a9543fcaf6b361fe4cd959b28ad60512c82b3
Signed-off-by: Marissa Wall <marissaw@google.com>
Signed-off-by: celtare21 <celtare21@gmail.com>
(cherry picked from commit d038b88a6eafa53cddbe90ba5dfe30e59c2533ef)