msm: kgsl: Take a refcount only for memtype ION in imported_mem_show
In imported_mem_show, we are only interested in entries of memory
type ION. Take a reference count only on ION entries and only if
they are not marked as pending free. This would avoid the overhead
of taking and putting a reference count on unneeded entries which
are generally much higher in number than ION entries.

Change-Id: I036d49b082f352e00333ed9212a098a4510abfe1
Signed-off-by: Deepak Kumar <dkumar@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>
(cherry picked from commit e722287d445bf7237f68f9814366f485430bef49)