msm: thermal: simple: Never raise max freq above unthrottled max
If the max frequency is set below the standard max, there is a
possibility of msm-thermal-simple raising the frequency in an attempt to
throttle if the target throttling zone has a higher max frequency. This
will be an issue when we unlock the max frequency setting to userspace.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: Oktapra Amtono <oktapra.amtono@gmail.com>
(cherry picked from commit 12fc59fe2b9ad468324cedf3beee53376397e8c9)