ARM: dts: sdm660-gpu: Add quirk qcom,secure_pt_early_alloc
The quirk is necessary for the GPU to initialize the secure
pagetable earlier, so that the hyp allocations will succeed.
This seems to be necessary only when hyp_secure_alloc is
enabled and only on Adreno A5xx.

Signed-off-by: YousefAlgadri <yusufgadrie@gmail.com>
Signed-off-by: Yasir-Siddiqui <www.mohammad.yasir.s@gmail.com>
Signed-off-by: Rizky Benggolo <travarilo@gmail.com>
(cherry picked from commit 6f4b5c1b22919e8e0e1f638333409c56ecc871ed)