sched: Handle remote wakeups on remote CPUs
When a task wakes up, and the current CPU does not share a cache with
the origin CPU, queue the task on the origin CPU in an attempt to take
advantage of cache locality. Fallback to standard wakeup handling if the
current CPU and the origin CPU share a cache.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
(cherry picked from commit a13cfa87b43b2a75b5cb8bdd8439cccd66bc6cad)