drm: Boost DDR bus upon running an atomic ioctl
In order to reduce jank, request a DDR bus boost whenever an atomic
ioctl is sent, which is what our diplay HAL uses to commit frames to the
display. The boost should be sufficient enough to render the display's
maximum FPS without any dropped frames when there is no significant
external source of load.

Signed-off-by: Crash-Overlord <el_jose47@live.com>